- split out sharepoint site collection
