+++ main.js
@@ -17,6 +17,7 @@ app.set('secretKey', process.env.SECRET_KEY || 'i-love-sharepoint');
 
 // setup api as /v1/api
 app.use('/v1/api', require('./routes/api'));
+app.use('/v1/api/site', require('./routes/site')); // sharepoint site collection
 app.use('/v1/api/employee', require('./routes/employee'));
 app.use('/v1/api/course', require('./routes/course'));
 app.use('/v1/api/video', require('./routes/video'));
